Change type of active_students on Course

+++ db/migrate/20161108142353_create_courses.rb
@@ -5,7 +5,7 @@ class CreateCourses < ActiveRecord::Migration
       t.text :description
       t.text :content
       t.string :duration
-      t.decimal :active_students
+      t.integer :active_students
       t.float :price
 
       t.belongs_to :school, index: true
